feat: add grpc

+++ internal/domain/car.go
@@ -3,9 +3,9 @@ package domain
 import "context"
 
 type CarRepository interface {
-	Find(ctx context.Context, id int) (*Car, error)
+	Find(ctx context.Context, id int64) (*Car, error)
 }
 
 type Car struct {
-	ID int
+	ID int64
 }
